Изменил положние функций в модулях
