+++ js/main.js
@@ -1,34 +1,51 @@
 import {renderElements} from "./util.js";
 import {hotelsInfo} from "./data.js";
-import {
-  onAdformInputCapacityChange,
-  onTypeInputChange,
-  onTimeinInputChange,
-  onTimeoutInputChange,
-  toggleFormElementsState
-} from "./form.js";
+import {inputsListeners, toggleFormElementsState, fillAdresInput} from "./form.js";
 import {createPin, onMapPinKeydown} from "./pin.js";
+
+const СontrolButtons = {
+  LEFTMOUSEBTN: 0,
+  ENTER: `Enter`,
+  ESCAPE: `Escape`,
+};
 
 const map = document.querySelector(`.map`);
 const adformElement = document.querySelector(`.ad-form`);
-const adformCapacityInput = adformElement.querySelector(`#capacity`);
-const adformTypeInput = adformElement.querySelector(`#type`);
-const adformTimeinInput = adformElement.querySelector(`#timein`);
-const adformTimeoutInput = adformElement.querySelector(`#timeout`);
 const filtersFormElement = document.querySelector(`.map__filters`);
 const pinsElement = document.querySelector(`.map__pins`);
+const mainPinElement = map.querySelector(`.map__pin--main`);
 
 const removeInactiveState = () => {
   map.classList.remove(`map--faded`);
   adformElement.classList.remove(`ad-form--disabled`);
   toggleFormElementsState(adformElement.children, false);
   toggleFormElementsState(filtersFormElement.children, false);
-  adformCapacityInput.addEventListener(`change`, onAdformInputCapacityChange);
-  adformTypeInput.addEventListener(`change`, onTypeInputChange);
-  adformTimeinInput.addEventListener(`change`, onTimeinInputChange);
-  adformTimeoutInput.addEventListener(`change`, onTimeoutInputChange);
+  inputsListeners();
   renderElements(hotelsInfo, pinsElement, createPin);
   map.addEventListener(`keydown`, onMapPinKeydown);
 };
 
-export {removeInactiveState};
+
+const onMainpinMousedown = (evt) => {
+  if (evt.button === СontrolButtons.LEFTMOUSEBTN) {
+    removeInactiveState();
+    fillAdresInput();
+    changeMainpinEventsState(false);
+  }
+};
+
+const onMainpinKeydown = (evt) => {
+  if (evt.key === СontrolButtons.ENTER) {
+    removeInactiveState();
+    fillAdresInput();
+    changeMainpinEventsState(false);
+  }
+};
+
+const changeMainpinEventsState = (type) => {
+  const method = type ? `addEventListener` : `removeEventListener`;
+  mainPinElement[method](`mousedown`, onMainpinMousedown);
+  mainPinElement[method](`keydown`, onMainpinKeydown);
+};
+
+changeMainpinEventsState(true);
